--- src/test/java/com/courtcanva/ccfranchise/utils/OpenOrderTestHelper.java
@@ -1,10 +1,0 @@
-package com.courtcanva.ccfranchise.utils;
-
-import com.courtcanva.ccfranchise.dtos.IdDto;
-import com.courtcanva.ccfranchise.dtos.suburbs.SuburbListPostDto;
-
-public class OpenOrderTestHelper {
-    public static IdDto createIdDto(Long mockFranchiseeId) {
-        return new IdDto(mockFranchiseeId);
-    }
-}
